Custom Exceptions and Testing

+++ src/main/java/com/parga/messagingapp/user/UserService.java
@@ -1,6 +1,8 @@
 package com.parga.messagingapp.user;
 
 import com.google.common.collect.Lists;
+import com.parga.messagingapp.DTO.UpdateUserDTO;
+import com.parga.messagingapp.exception.FieldNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -42,28 +44,53 @@ public class UserService {
         return Lists.newArrayList(iterable);
     }
 
-    public User getUserById(String id){
+    public User getUserById(String id) throws Exception{
+
+        if( id == null || id.isEmpty() )
+            throw new FieldNotFoundException("id must not be null or empty");
+
         Optional<User> optionalUser = userRepository.findById(id);
         return optionalUser.get();
     }
 
-    public User createUser(User user){
+    public User createUser(User user) throws Exception{
+
+        if( user.getUsername() == null || user.getUsername().isEmpty())
+            throw new FieldNotFoundException("Username must not be null or empty");
+
+        if( user.getPassword() == null || user.getPassword().isEmpty())
+            throw new FieldNotFoundException("Password must not be null or empty");
+
         return  userRepository.insert(user);
     }
 
-    public User updateUser(String id, User user) {
+    public User updateUser(String id, UpdateUserDTO user) throws Exception{
+
+        if( id == null || id.isEmpty() )
+            throw new FieldNotFoundException("id must not be null or empty");
+
+        if (user.getUsername() == null || user.getUsername().isEmpty())
+            throw new FieldNotFoundException("username must not be null or empty");
 
         User userToUpdate = userRepository.findById(id).get();
         if( userToUpdate == null )
             return null;
 
         userToUpdate.setUsername(user.getUsername());
-        userToUpdate.setPassword(user.getPassword());
 
         return userRepository.save(userToUpdate);
     }
 
-    public void deleteUser(String id) {
+    public void deleteUser(String id) throws Exception{
+
+        if( id == null || id.isEmpty() )
+            throw new FieldNotFoundException("id must not be null or empty");
+
+        User user = userRepository.findById(id).orElse(null);
+
+        if(user == null)
+            throw new FieldNotFoundException("user must not be null or empty");
+
         userRepository.deleteById(id);
     }
 }
